Add food and shopping list apps, added food categories
I am planning on importing the food information from bring which will
give me a great start (and quite possibley I won't need much more than
that). Plan on bringing in more stuff as well, so a nightly? script that
grabs the data might be a good idea.

If I bring in the food data then I can easily add stuff to bring as the
foods have their related id names (in german I think?) which are needed
to add the correct food to the bring API. Solves a pretty big issue I
had with working with their API.