Rename food to products, basic bring import script
The bring API is not just food items but also had a bunch of other
stuff. While I could specify specific categories I pretty much decided I
wasn't that fussed and that all products are welcome here xD.

The import script is very basic, doesn't delete before it imports and
doesn't check for already imported items. I want to be able to run this
script nightly to pick up any new data so going to make it so that it
picks up changes only.