Add bring id and week day to shopping lists
The idea is that the shopping lists represent the lists on bring. Meals
have ingredience and a date, when it is saved there will be a check to
see if the meal is in the next week, if it is it will assign it to the
correct shopping list for the day in Bring.